20260410--ai总体优化，销售模块优化第一步

+++ platform-modules/module-decision-support/decision-support-service/src/main/java/com/cjx/decision/dto/dashboard/RawCollection.java
@@ -2,10 +2,12 @@ package com.cjx.decision.dto.dashboard;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RawCollection {
     private String collectionAmount;
-    private Double collectionRate;
-    private Double gapValue;
-    private Double monthGoal;
+    private BigDecimal collectionRate;
+    private BigDecimal gapValue;
+    private BigDecimal monthGoal;
 }
