+++ platform-modules/module-decision-support/decision-support-service/src/main/java/com/cjx/decision/dto/dashboard/RawSalesMetric.java
@@ -2,12 +2,14 @@ package com.cjx.decision.dto.dashboard;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RawSalesMetric {
     private String metricName;
     private String displayValue;
     private String type;
-    private Double budgetRate;
-    private Double gapValue;
-    private Double monthGoal;
+    private BigDecimal budgetRate;
+    private BigDecimal gapValue;
+    private BigDecimal monthGoal;
 }
